Tested and working in sim. Applied a value of 45,000 to the conductivity and am seeing ~55,600 for specific conductivity in liaison
