+++ src/bin/drivers/atlas_scientific_ezo_ec/app.cpp
@@ -36,7 +36,6 @@
 #include "jaiabot/messages/moos.pb.h"
 #include "jaiabot/messages/sensor/pressure_temperature.pb.h"
 #include "jaiabot/messages/sensor/salinity.pb.h"
-#include "jaiabot/utils/specific_conductivity.h"
 
 using goby::glog;
 namespace si = boost::units::si;
@@ -78,6 +77,16 @@ int main(int argc, char* argv[])
 {
     return goby::run<jaiabot::apps::AtlasSalinityPublisher>(
         goby::middleware::ProtobufConfigurator<config::AtlasSalinityPublisher>(argc, argv));
+}
+
+double calculate_specific_conductivity(double conductivity, double temperature)
+{
+    double temperature_coefficient = 0.0191;
+    double EC_25;
+
+    EC_25 = conductivity / (1 + temperature_coefficient * (temperature - 25));
+
+    return EC_25;
 }
 
 // Main thread
