+++ src/bin/sensors/drivers/atlas_scientific__oem_ec.h
@@ -26,7 +26,9 @@
 #include "config.pb.h"
 #include "jaiabot/messages/health.pb.h"
 #include "jaiabot/messages/sensor/atlas_scientific__oem_ec.pb.h"
+#include "jaiabot/messages/sensor/pressure_temperature.pb.h"
 #include "jaiabot/messages/sensor/sensor_core.pb.h"
+
 #include <goby/zeromq/application/multi_thread.h>
 
 namespace jaiabot
@@ -50,6 +52,7 @@ class AtlasScientificOEMECDriver
     int32_t sample_rate_{10};
     int32_t report_timeout_{20};
     int32_t resend_cfg_timeout_{20};
+    jaiabot::protobuf::PressureTemperatureData last_pressure_temperature_data_;
 };
 
 } // namespace apps
